# docs/api/quant-radar-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Surge Screener API
  version: 1.23.0-draft
  summary: Loopback-only public reads plus protected private operator state.
  description: |
    Contract for the current Phase 1 fixed-artifact reads plus bounded Phase 2
    reads for IV history, the institutional fund catalog, the manually maintained
    AI updates feed, the manually maintained schedule registry, a strict
    Options Flow and candidate feed projections, the strict Crypto Universe,
    Social Intelligence, Theme Flow, Money Flow, Sector Rotation, Daily Summary,
    Playbook Validation, Continuation Validation, COT report catalog/detail, and
    bounded Market Thesis validation/regime-history, Oversold Reversal
    validation, Reversal Radar validation, Knowledge Graph, Watchlist theme
    taxonomy, sector-to-theme drill, and Influencer roster public projections.
    Expected
    artifact failures are represented as HTTP 200 unavailable envelopes so a
    missing or half-written scheduled output cannot crash the frontend. Invalid
    request parameters use sanitized HTTP 422 problem details. Unexpected server
    defects remain HTTP 500.

    Public artifact routes intentionally have no authentication scheme because
    the server is restricted to 127.0.0.1 and exposes bounded non-sensitive
    projections. The private Industry Roles review-board routes additionally
    require a high-entropy Streamlit-to-FastAPI bearer credential. Mutations
...
servers:
  - url: http://127.0.0.1:8000
...
security: []
...
paths:
...
  /api/v1/market-context/theme-drill:
    get:
      tags: [Market Context]
      summary: Read the fixed sector-to-theme drill projection
      description: Projects only theme names and parent SPDR sector ETFs from the fixed basket curation; ticker membership and producer details are omitted.
      operationId: getThemeDrill
      responses:
        "200":
          $ref: "#/components/responses/ThemeDrillResponse"
        "500":
          $ref: "#/components/responses/InternalProblem"
...
components:
  securitySchemes:
    InternalServiceBearer:
      type: http
      scheme: bearer
      description: |
        High-entropy Streamlit-to-FastAPI service credential for protected
        loopback routes; not a human login or provider token.
  headers:
    NoStore:
      description: Prevent browsers and intermediaries from retaining artifact state.
      schema:
        type: string
        const: no-store
...
  responses:
...
    ThemeDrillResponse:
      description: Fixed public sector-to-theme drill state.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ThemeDrillEnvelope"
          examples:
            available:
              summary: A populated fixed public projection
              value:
                available: true
                reason: ok
                data:
                  sectors:
                    - etf: XLK
                      themes: [HBM, Advanced packaging]
                meta:
                  sourceId: market-context.theme-drill
            empty:
              summary: A valid empty public projection
              value:
                available: true
                reason: ok
                data:
                  sectors: []
                meta:
                  sourceId: market-context.theme-drill
            unavailable:
              summary: The fixed source is missing
              value:
                available: false
                reason: missing
                meta:
                  sourceId: market-context.theme-drill
...
    InternalProblem:
      description: Unexpected programming or server failure.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: about:blank
            title: Internal Server Error
            status: 500
  schemas:
...
    AvailableArtifactBase:
      type: object
      required: [available, reason, meta]
      properties:
        available:
          type: boolean
          const: true
        reason:
          type: string
          const: ok
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
...
    AvailableThemeDrill:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/AvailableArtifactBase"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/ThemeDrillData"
...
    UnavailableArtifact:
      type: object
      additionalProperties: false
      required: [available, reason, data, meta]
      properties:
        available:
          type: boolean
          const: false
        reason:
          $ref: "#/components/schemas/UnavailableReason"
        data:
          type: "null"
        meta:
          $ref: "#/components/schemas/ArtifactMeta"
    UnavailableReason:
      type: string
      enum:
        - missing
        - invalid_json
        - invalid_shape
        - unreadable
...
    ThemeDrillEnvelope:
      oneOf:
        - $ref: "#/components/schemas/AvailableThemeDrill"
        - $ref: "#/components/schemas/UnavailableArtifact"
...
    ThemeDrillData:
      type: object
      additionalProperties: false
      required: [sectors]
      properties:
        sectors:
          type: array
          maxItems: 11
          items:
            $ref: "#/components/schemas/ThemeDrillSector"
    ThemeDrillSector:
      type: object
      additionalProperties: false
      required: [etf, themes]
      properties:
        etf:
          type: string
          enum: [XLB, XLC, XLE, XLF, XLI, XLK, XLP, XLRE, XLU, XLV, XLY]
        themes:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 200
            pattern: '^[^\x00-\x08\x0b\x0c\x0e-\x1f\x7f]+$'
...
    ArtifactMeta:
      type: object
      additionalProperties: false
      required: [sourceId, asOf, generatedAt]
      properties:
        sourceId:
          type: string
          description: Stable registry identifier; never a filesystem path.
          pattern: "^[a-z0-9]+(?:[.-][a-z0-9]+)*$"
        asOf:
          type: [string, "null"]
          format: date
        generatedAt:
          type: [string, "null"]
          format: date-time
    Problem:
      type: object
      description: RFC 9457 problem details for unexpected server failures.
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference